// src/app/shared/models/system-users-address.ts
export class UsersAddress {
    name: string;
    street: string;
    number: string;
    complement: string;
    neighborhood: string;
    zip_code: string;
    city: string;
    state: string;

    constructor(
        name: string,
        street: string,
        number: string,
        complement: string,
        neighborhood: string,
        zip_code: string,
        city: string,
        state: string
    ) { }
}